v1.6-SNAPSHOT
Fix reporting of results when testplan contains multiple runs

We look for system property testRail.runConfig for the name of the configuration that was executed in test run. This along with automationId of test is used to look up the test run for which the result has to be reported.