v2.2
added more information when test result is blocked. This would happen when test is skipped which would happen if the configuration method failed. If that's the case, we try and store this information in ThreadLocal (in case of parallel execution) and then send it over with the result of the test.